[Feat] In-Progress Scaling
Calculate marks for on-going modules by scaling the overall mark by the amount completed - allowing for an accurate indication of user's performance.

This also allows for some assessments to not be inputted until they are completed, without leading to a very low percentage.

Next, need to add checking for assessments in the future that are causing NaN values